gh-pages: Add files for GitHub Pages
Ideally these files should not be here - they are specific to one static
site builder we use at the moment, Jekyll + GitHub Pages. But since they
are so small I add them to `master` for convenience. It allows to just
build the site from `master` without the additional step of adding these
files.